Tests now works
