+++ src/test/java/pl/edu/pwr/student/Gates/BasicGates/MultipleInput/ANDTest.java
@@ -1,17 +1,20 @@
-package pl.edu.pwr.student.Gates.BasicGates;
+package pl.edu.pwr.student.Gates.BasicGates.MultipleInput;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import pl.edu.pwr.student.IO.Input.Switch;
 import pl.edu.pwr.student.IO.Output.LED;
 
 import static org.junit.jupiter.api.Assertions.*;
+import static pl.edu.pwr.student.Simulation.simWait;
 
-class NORTest {
+class ANDTest {
     private Switch FirstInput = new Switch();
     private Switch SecondInput = new Switch();
     private pl.edu.pwr.student.IO.Output.LED LED = new LED("led",100);
-    private NOR gate = new NOR();
+    private AND gate = new AND();
 
+    @BeforeEach
     protected void setUp() {
         if(FirstInput.getState())
             FirstInput.toggle();
@@ -21,48 +24,52 @@ class NORTest {
         FirstInput.connection(gate);
         SecondInput.connection(gate);
         gate.connection(LED);
+        simWait(50);
     }
 
     /*
-     *Test NOR gate input 0 and 0
+     *Test AND gate input 0 and 0
      */
     @Test
     public void testTruthTableCase1() {
-        assertTrue(gate.getState());
+        assertFalse(gate.getState());
     }
 
     /*
-     *Test NOR gate input 1 and 0
+     *Test AND gate input 1 and 0
      */
     @Test
     public void testTruthTableCase2() {
         FirstInput.toggle();
+        simWait(50);
         assertFalse(gate.getState());
     }
     /*
-     * Test NOR gate input 0 and 1
+     * Test AND gate input 0 and 1
      */
     @Test
     public void testTruthTableCase3() {
         SecondInput.toggle();
+        simWait(50);
         assertFalse(gate.getState());
     }
     /*
-     *Test NOR gate input 1 and 1
+     *Test AND gate input 1 and 1
      */
     @Test
     public void testTruthTableCase4() {
         FirstInput.toggle();
         SecondInput.toggle();
+        simWait(50);
 
-        assertFalse(gate.getState());
+        assertTrue(gate.getState());
     }
     /*
      * Test connection with two inputs
      */
     @Test
     public void testGateHasInputs() {
-        assertTrue(gate.hasInputs());
+        assertFalse(gate.getInputs().isEmpty());
     }
     /*
      * Test disconnect inputs
@@ -70,7 +77,8 @@ class NORTest {
     @Test
     public void testDisconnectInputs() {
         gate.disconnectInputs();
-        assertFalse(gate.hasInputs());
+        simWait(50);
+        assertTrue(gate.getInputs().isEmpty());
     }
     /*
      * Test full disconnection
@@ -78,7 +86,8 @@ class NORTest {
     @Test
     public void testFullDisconnect() {
         gate.fullDisconnect();
-        assertFalse(gate.hasInputs());
+        simWait(50);
+        assertTrue(gate.getInputs().isEmpty());
     }
     /*
      * Test disconnect outputs
@@ -86,6 +95,7 @@ class NORTest {
     @Test
     public void testDisconnectOutputs() {
         gate.disconnectOutputs();
+        simWait(50);
         assertEquals(LED.getOutputs().size(),0);
     }
 }
