+++ src/test/java/pl/edu/pwr/student/Gates/BasicGates/MultipleInput/NANDTest.java
@@ -1,91 +1,107 @@
-package pl.edu.pwr.student.Gates.BasicGates;
+package pl.edu.pwr.student.Gates.BasicGates.MultipleInput;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import pl.edu.pwr.student.IO.Input.Switch;
 import pl.edu.pwr.student.IO.Output.LED;
 
 import static org.junit.jupiter.api.Assertions.*;
+import static pl.edu.pwr.student.Simulation.simWait;
 
-class ANDTest {
+class NANDTest {
     private Switch FirstInput = new Switch();
     private Switch SecondInput = new Switch();
-    private pl.edu.pwr.student.IO.Output.LED LED = new LED("led",100);
-    private AND gate = new AND();
+    private pl.edu.pwr.student.IO.Output.LED LED = new LED("led", 100);
+    private NAND gate = new NAND();
 
+    @BeforeEach
     protected void setUp() {
-        if(FirstInput.getState())
+        if (FirstInput.getState())
             FirstInput.toggle();
-        if(SecondInput.getState())
+        if (SecondInput.getState())
             SecondInput.toggle();
 
         FirstInput.connection(gate);
         SecondInput.connection(gate);
         gate.connection(LED);
+        simWait(50);
     }
 
     /*
-     *Test AND gate input 0 and 0
+     *Test NAND gate input 0 and 0
      */
     @Test
     public void testTruthTableCase1() {
-        assertFalse(gate.getState());
+        assertTrue(gate.getState());
     }
 
     /*
-     *Test AND gate input 1 and 0
+     *Test NAND gate input 1 and 0
      */
     @Test
     public void testTruthTableCase2() {
         FirstInput.toggle();
-        assertFalse(gate.getState());
+        simWait(50);
+        assertTrue(gate.getState());
     }
+
     /*
-     * Test AND gate input 0 and 1
+     * Test NAND gate input 0 and 1
      */
     @Test
     public void testTruthTableCase3() {
         SecondInput.toggle();
-        assertFalse(gate.getState());
+        simWait(50);
+        assertTrue(gate.getState());
     }
+
     /*
-     *Test AND gate input 1 and 1
+     *Test NAND gate input 1 and 1
      */
     @Test
     public void testTruthTableCase4() {
         FirstInput.toggle();
         SecondInput.toggle();
+        simWait(50);
 
-        assertTrue(gate.getState());
+        assertFalse(gate.getState());
     }
+
     /*
      * Test connection with two inputs
      */
     @Test
     public void testGateHasInputs() {
-        assertTrue(gate.hasInputs());
+        assertFalse(gate.getInputs().isEmpty());
     }
+
     /*
      * Test disconnect inputs
      */
     @Test
     public void testDisconnectInputs() {
         gate.disconnectInputs();
-        assertFalse(gate.hasInputs());
+        simWait(50);
+        assertTrue(gate.getInputs().isEmpty());
     }
+
     /*
      * Test full disconnection
      */
     @Test
     public void testFullDisconnect() {
         gate.fullDisconnect();
-        assertFalse(gate.hasInputs());
+        simWait(50);
+        assertTrue(gate.getInputs().isEmpty());
     }
+
     /*
      * Test disconnect outputs
      */
     @Test
     public void testDisconnectOutputs() {
         gate.disconnectOutputs();
-        assertEquals(LED.getOutputs().size(),0);
+        simWait(50);
+        assertEquals(LED.getOutputs().size(), 0);
     }
 }
