+++ src/test/java/pl/edu/pwr/student/Gates/BasicGates/MultipleInput/ORTest.java
@@ -1,17 +1,20 @@
-package pl.edu.pwr.student.Gates.BasicGates;
+package pl.edu.pwr.student.Gates.BasicGates.MultipleInput;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import pl.edu.pwr.student.IO.Input.Switch;
 import pl.edu.pwr.student.IO.Output.LED;
 
 import static org.junit.jupiter.api.Assertions.*;
+import static pl.edu.pwr.student.Simulation.simWait;
 
-class XORTest {
+class ORTest {
     private Switch FirstInput = new Switch();
     private Switch SecondInput = new Switch();
     private pl.edu.pwr.student.IO.Output.LED LED = new LED("led",100);
-    private XOR gate = new XOR();
+    private OR gate = new OR();
 
+    @BeforeEach
     protected void setUp() {
         if(FirstInput.getState())
             FirstInput.toggle();
@@ -21,10 +24,11 @@ class XORTest {
         FirstInput.connection(gate);
         SecondInput.connection(gate);
         gate.connection(LED);
+        simWait(50);
     }
 
     /*
-     *Test XOR gate input 0 and 0
+     *Test OR gate input 0 and 0
      */
     @Test
     public void testTruthTableCase1() {
@@ -32,37 +36,40 @@ class XORTest {
     }
 
     /*
-     *Test XOR gate input 1 and 0
+     *Test OR gate input 1 and 0
      */
     @Test
     public void testTruthTableCase2() {
         FirstInput.toggle();
+        simWait(50);
         assertTrue(gate.getState());
     }
     /*
-     * Test XOR gate input 0 and 1
+     * Test OR gate input 0 and 1
      */
     @Test
     public void testTruthTableCase3() {
         SecondInput.toggle();
+        simWait(50);
         assertTrue(gate.getState());
     }
     /*
-     *Test XOR gate input 1 and 1
+     *Test OR gate input 1 and 1
      */
     @Test
     public void testTruthTableCase4() {
         FirstInput.toggle();
         SecondInput.toggle();
+        simWait(50);
 
-        assertFalse(gate.getState());
+        assertTrue(gate.getState());
     }
     /*
      * Test connection with two inputs
      */
     @Test
     public void testGateHasInputs() {
-        assertTrue(gate.hasInputs());
+        assertFalse(gate.getInputs().isEmpty());
     }
     /*
      * Test disconnect inputs
@@ -70,7 +77,8 @@ class XORTest {
     @Test
     public void testDisconnectInputs() {
         gate.disconnectInputs();
-        assertFalse(gate.hasInputs());
+        simWait(50);
+        assertTrue(gate.getInputs().isEmpty());
     }
     /*
      * Test full disconnection
@@ -78,7 +86,8 @@ class XORTest {
     @Test
     public void testFullDisconnect() {
         gate.fullDisconnect();
-        assertFalse(gate.hasInputs());
+        simWait(50);
+        assertTrue(gate.getInputs().isEmpty());
     }
     /*
      * Test disconnect outputs
@@ -86,6 +95,7 @@ class XORTest {
     @Test
     public void testDisconnectOutputs() {
         gate.disconnectOutputs();
+        simWait(50);
         assertEquals(LED.getOutputs().size(),0);
     }
 }
