+++ src/test/java/pl/edu/pwr/student/Gates/BasicGates/MultipleInput/XORTest.java
@@ -1,38 +1,67 @@
-package pl.edu.pwr.student.Gates.BasicGates;
+package pl.edu.pwr.student.Gates.BasicGates.MultipleInput;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import pl.edu.pwr.student.IO.Input.Switch;
 import pl.edu.pwr.student.IO.Output.LED;
 
 import static org.junit.jupiter.api.Assertions.*;
+import static pl.edu.pwr.student.Simulation.simWait;
 
-class NOTTest {
+class XORTest {
     private Switch FirstInput = new Switch();
+    private Switch SecondInput = new Switch();
     private pl.edu.pwr.student.IO.Output.LED LED = new LED("led",100);
-    private NOT gate = new NOT();
+    private XOR gate = new XOR();
 
+    @BeforeEach
     protected void setUp() {
         if(FirstInput.getState())
             FirstInput.toggle();
+        if(SecondInput.getState())
+            SecondInput.toggle();
 
         FirstInput.connection(gate);
+        SecondInput.connection(gate);
         gate.connection(LED);
+        simWait(50);
     }
 
     /*
-     *Test NOT gate input 0
+     *Test XOR gate input 0 and 0
      */
     @Test
     public void testTruthTableCase1() {
-        assertTrue(gate.getState());
+        assertFalse(gate.getState());
     }
 
     /*
-     *Test NOT gate input 1
+     *Test XOR gate input 1 and 0
      */
     @Test
     public void testTruthTableCase2() {
         FirstInput.toggle();
+        simWait(50);
+        assertTrue(gate.getState());
+    }
+    /*
+     * Test XOR gate input 0 and 1
+     */
+    @Test
+    public void testTruthTableCase3() {
+        SecondInput.toggle();
+        simWait(50);
+        assertTrue(gate.getState());
+    }
+    /*
+     *Test XOR gate input 1 and 1
+     */
+    @Test
+    public void testTruthTableCase4() {
+        FirstInput.toggle();
+        SecondInput.toggle();
+        simWait(50);
+
         assertFalse(gate.getState());
     }
     /*
@@ -40,7 +69,7 @@ class NOTTest {
      */
     @Test
     public void testGateHasInputs() {
-        assertTrue(gate.hasInputs());
+        assertFalse(gate.getInputs().isEmpty());
     }
     /*
      * Test disconnect inputs
@@ -48,7 +77,8 @@ class NOTTest {
     @Test
     public void testDisconnectInputs() {
         gate.disconnectInputs();
-        assertFalse(gate.hasInputs());
+        simWait(50);
+        assertTrue(gate.getInputs().isEmpty());
     }
     /*
      * Test full disconnection
@@ -56,7 +86,8 @@ class NOTTest {
     @Test
     public void testFullDisconnect() {
         gate.fullDisconnect();
-        assertFalse(gate.hasInputs());
+        simWait(50);
+        assertTrue(gate.getInputs().isEmpty());
     }
     /*
      * Test disconnect outputs
@@ -64,6 +95,7 @@ class NOTTest {
     @Test
     public void testDisconnectOutputs() {
         gate.disconnectOutputs();
+        simWait(50);
         assertEquals(LED.getOutputs().size(),0);
     }
 }
