+++ src/test/java/pl/edu/pwr/student/Gates/BasicGates/SingleInput/NOTTest.java
@@ -1,30 +1,30 @@
-package pl.edu.pwr.student.Gates.BasicGates;
+package pl.edu.pwr.student.Gates.BasicGates.SingleInput;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import pl.edu.pwr.student.IO.Input.Switch;
 import pl.edu.pwr.student.IO.Output.LED;
 
 import static org.junit.jupiter.api.Assertions.*;
+import static pl.edu.pwr.student.Simulation.simWait;
 
-class XNORTest {
+class NOTTest {
     private Switch FirstInput = new Switch();
-    private Switch SecondInput = new Switch();
-    private pl.edu.pwr.student.IO.Output.LED LED = new LED("led",100);
-    private XNOR gate = new XNOR();
+    private LED LED = new LED("led",100);
+    private NOT gate = new NOT();
 
+    @BeforeEach
     protected void setUp() {
-        FirstInput.connection(gate);
-        SecondInput.connection(gate);
-        gate.connection(LED);
-
         if(FirstInput.getState())
             FirstInput.toggle();
-        if(SecondInput.getState())
-            SecondInput.toggle();
+
+        FirstInput.connection(gate);
+        gate.connection(LED);
+        simWait(50);
     }
 
     /*
-     *Test XNOR gate input 0 and 0
+     *Test NOT gate input 0
      */
     @Test
     public void testTruthTableCase1() {
@@ -32,30 +32,13 @@ class XNORTest {
     }
 
     /*
-     *Test XNOR gate input 1 and 0
+     *Test NOT gate input 1
      */
     @Test
     public void testTruthTableCase2() {
         FirstInput.toggle();
+        simWait(50);
         assertFalse(gate.getState());
-    }
-    /*
-     * Test XNOR gate input 0 and 1
-     */
-    @Test
-    public void testTruthTableCase3() {
-        SecondInput.toggle();
-        assertFalse(gate.getState());
-    }
-    /*
-     *Test XNOR gate input 1 and 1
-     */
-    @Test
-    public void testTruthTableCase4() {
-        FirstInput.toggle();
-        SecondInput.toggle();
-
-        assertTrue(gate.getState());
     }
     /*
      * Test connection with two inputs
@@ -70,6 +53,7 @@ class XNORTest {
     @Test
     public void testDisconnectInputs() {
         gate.disconnectInputs();
+        simWait(50);
         assertFalse(gate.hasInputs());
     }
     /*
@@ -78,6 +62,7 @@ class XNORTest {
     @Test
     public void testFullDisconnect() {
         gate.fullDisconnect();
+        simWait(50);
         assertFalse(gate.hasInputs());
     }
     /*
@@ -86,6 +71,7 @@ class XNORTest {
     @Test
     public void testDisconnectOutputs() {
         gate.disconnectOutputs();
+        simWait(50);
         assertEquals(LED.getOutputs().size(),0);
     }
 }
