+++ src/test/java/pl/edu/pwr/student/IO/Output/DebugLEDTest.java
@@ -1,14 +1,16 @@
 package pl.edu.pwr.student.IO.Output;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import pl.edu.pwr.student.IO.Input.Switch;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class LEDTest {
+class DebugLEDTest {
     private Switch FirstInput = new Switch();
-    private LED LED = new LED("LED",100);
+    private DebugLED LED = new DebugLED("LED",100);
 
+    @BeforeEach
     protected void setUp() {
         FirstInput.connection(LED);
     }
