+++ src/test/java/pl/edu/pwr/student/IO/Output/SpeakerTest.java
@@ -1,5 +1,6 @@
 package pl.edu.pwr.student.IO.Output;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import pl.edu.pwr.student.IO.Input.Switch;
 
@@ -9,6 +10,7 @@ class SpeakerTest {
     private Switch FirstInput = new Switch();
     private Speaker Speaker = new Speaker();
 
+    @BeforeEach
     protected void setUp() {
         FirstInput.connection(Speaker);
     }
